Criar página para apagar a conta

+++ src/main/route.py
@@ -1,5 +1,6 @@
-from flask import Blueprint, render_template, request, redirect, url_for
+from flask import Blueprint, render_template, request, redirect, url_for, flash
 from flask_login import logout_user, login_required, current_user
+from werkzeug.security import check_password_hash
 from database.model import db, User
 
 
@@ -11,14 +12,19 @@ def index():
     return render_template('main/index.html', tasks=(current_user.is_authenticated and current_user.tasks))
 
 
-@main_bp.route('/delete_account/<int:user_id>', methods=['POST'])
+@main_bp.route('/delete_account', methods=['GET', 'POST'])
 @login_required
-def delete_account(user_id):
+def delete_account():
     if request.method == 'POST':
-        user = db.session.get(User, user_id)
-        for task in user.tasks:
-            db.session.delete(task)
-        logout_user()
-        db.session.delete(user)
-        db.session.commit()
-    return redirect(url_for('main.index'))
+        password = request.form['password']
+        user = db.session.get(User, current_user.id)
+
+        if check_password_hash(user.password, password):
+            for task in user.tasks:
+                db.session.delete(task)
+            logout_user()
+            db.session.delete(user)
+            db.session.commit()
+            return redirect(url_for('main.index'))
+        flash('Incorrect Password', category='error')
+    return render_template('main/delete_account.html')
